+++ src/pages/LoginPage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import LoginForm from '../forms/LoginForm';
+import { LoginForm } from '../components';
 import { useTheme } from '../hooks/useTheme';
 
 export default function LoginPage() {
